feat: Add new inventory files for A4 and A5 formats
Added new inventory files for A4 and A5 formats. These files contain the list of available products and their corresponding IDs and quantities. The A4 inventory file has 37 items, while the A5 inventory file has 45 items.

The inventory files will be used to generate QR codes for each product, which will be included in the product templates.

This commit addresses the need to update the inventory and ensure that the correct information is used for generating QR codes.